Script to find woring apr.

+++ setup.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-import os, sys
+import os, sys, time
 
 # TODO: place magical installation code here
 
@@ -11,15 +11,23 @@ SETUP_DEBUG = True
 #  GETAPR_LIST  #
 #################
 def getAPR_list() :
-  os.system( 'find / -name "apr_file_io.h" | grep -v "Permission denied" > out.txt' )
+  cmd = 'find / -name "apr_file_io.h" | grep -v "Permission denied" > out.txt'
+  print "Finding Apache Runtime library using command: " + cmd
+  time.sleep(5) # message to user
+  os.system( cmd )
   fo = open( "out.txt", "r" )
 
   pathList = []
   for path in fo :
+    path = path.strip()
+    path_split = path.split( "/" )
+    path_split = path_split[:len(path_split)-1]
+    path       = "/".join( path_split )
     pathList.append( path )
 
   os.system( 'rm out.txt' )
 
+  return pathList
 
 ##############################
 #  MAIN THREAD OF EXECUTION  #
@@ -27,26 +35,52 @@ def getAPR_list() :
 
 print "Running pyLDFI setup with : \n" + str(sys.argv)
 
+# run make for c4
 # find candidate apr locations
 apr_path_cands = getAPR_list()
 
 # set correct apr location
+flag    = True
 for path in apr_path_cands :
+  # http://stackoverflow.com/questions/4710067/deleting-a-specific-line-in-a-file-python
+  # protect against multiple runs of setup
+  f = open( C4_FINDAPR_PATH, "r+" )
+  d = f.readlines()
+  f.seek(0)
+  for i in d:
+    if not "set(APR_INCLUDES" in i :
+      f.write(i)
+  f.truncate()
+  f.close()
 
-  NOT_INSTALLED = True
-  while NOT_INSTALLED == True :
-    # set one of the candidate APR paths
-    os.system( "sed 's/" + '_apr_invoke(APR_VERSION   ""        --version)' + "/g'" + " " + C4_FINDAPR_PATH + " > " + '_apr_invoke(APR_VERSION   ""        --version)\nset(APR_INCLUDES "' + path + '")'  )
-    os.system( "make" )
- 
+  # set one of the candidate APR paths
+  newCmd = 'set(APR_INCLUDES "' + path + '")'
+  #cmd = "echo '" + newCmd + "' | cat - " + C4_FINDAPR_PATH + " > temp && mv temp " + C4_FINDAPR_PATH
+  cmd = "(head -48 " + C4_FINDAPR_PATH + "; " + "echo '" + newCmd + "'; " + "tail -n +49 " + C4_FINDAPR_PATH + ")" + " > temp ; mv temp " + C4_FINDAPR_PATH + ";"
+  os.system( cmd )
+  os.system( "make c4" )
+
+  try :
     fo = open( "./c4_out.txt", "r" )
     for line in fo :
-      if ">>> C4 Installation SUCCESSFUL <<<" in line :
-        NOT_INSTALLED = False
-    fo.close()
-    os.system( "rm ./c4_out.txt" ) # clean up
+      line = line.strip()
+      if "error generated." in line :
+        print "failed path = " + path
+        flag = False
+  except IOError :
+    print "./c4_out.txt does not exist"
+
+  # found a valid apr library
+  if flag :
+    break
+
+  fo.close()
+  os.system( "rm ./c4_out.txt" ) # clean up
 
   print "Using APR path : " + path
+
+# run make for everything else
+os.system( "make" )
 
 print "... Done installing C4 Datalog evaluator"
